ZoomToRectangle implemented Beholder now uses double (needed some fixes in other classes, too).
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@304 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ graphics/include/vislib/graphicstypes.h
@@ -18,7 +18,7 @@ namespace vislib {
 namespace graphics {
 
     /** This type is used for values in scene space. */
-    typedef float SceneSpaceType;
+    typedef double SceneSpaceType;
 
     /** 
      * This type is used for values in image space.
